+++ source/distributed/tria.inst.in
@@ -17,41 +17,6 @@
 
 for (deal_II_dimension : DIMENSIONS)
 {
-
-#   ifdef DEAL_II_WITH_P4EST
-
-    namespace internal
-    \{
-    namespace p4est
-    \{
-#if     deal_II_dimension > 1
-    template
-    void
-    init_quadrant_children<deal_II_dimension>
-    (const types<deal_II_dimension>::quadrant & p4est_cell,
-     types<deal_II_dimension>::quadrant (&p4est_children)[GeometryInfo<deal_II_dimension>::max_children_per_cell]);
-
-    template
-    void
-    init_coarse_quadrant<deal_II_dimension>
-    (types<deal_II_dimension>::quadrant & quad);
-
-    template
-    bool
-    quadrant_is_equal<deal_II_dimension>
-    (const types<deal_II_dimension>::quadrant & q1,
-     const types<deal_II_dimension>::quadrant & q2);
-
-    template
-    bool
-    quadrant_is_ancestor<deal_II_dimension>
-    (const types<deal_II_dimension>::quadrant & q1,
-     const types<deal_II_dimension>::quadrant & q2);
-#      endif
-    \}
-    \}
-#   endif // DEAL_II_WITH_P4EST
-
     namespace parallel
     \{
     namespace distributed
